Add ORD and ENUM xtypes at Py level

+++ ego/src/types.rs
@@ -97,8 +97,8 @@ pub enum XType {
     Cont(f64, f64),
     /// Integer variable in lower bound .. upper bound
     Int(i32, i32),
-    /// An Ordered variable in { int_1, int_2, ... int_n }
-    Ord(Vec<i32>),
+    /// An Ordered variable in { float_1, float_2, ... float_n }
+    Ord(Vec<f64>),
     /// An Enum variable in { str_1, str_2, ..., str_n }
     Enum(Vec<String>),
 }
